types added to navigation
following https://reactnavigation.org/docs/typescript
